fix: break up the utils to multiple files
