+++ src/authn/AuthData.ts
@@ -1,6 +1,6 @@
 import { authn as authnProto } from '@xmtp/proto'
 import Long from 'long'
-import { dateToNs } from '../utils'
+import { dateToNs } from '../utils/date'
 
 export default class AuthData implements authnProto.AuthData {
   walletAddr: string
